Move action-related logic + data models to their own folder
