+++ prime/src/components/borrow/ManageAccountWidget.tsx
@@ -14,7 +14,8 @@ import { ReactComponent as AlertTriangleIcon } from '../../assets/svg/alert_tria
 import { ReactComponent as CheckIcon } from '../../assets/svg/check_black.svg';
 import { ReactComponent as LoaderIcon } from '../../assets/svg/loader.svg';
 import { ReactComponent as PlusIcon } from '../../assets/svg/plus.svg';
-import { Action, ActionCardState, ActionID, TokenType, UniswapPosition } from '../../data/Actions';
+import { ActionID } from '../../data/actions/ActionID';
+import { Action, ActionCardState, TokenType, UniswapPosition } from '../../data/actions/Actions';
 import { RESPONSIVE_BREAKPOINT_SM, RESPONSIVE_BREAKPOINT_XS } from '../../data/constants/Breakpoints';
 import { UINT256_MAX } from '../../data/constants/Values';
 import { Assets, Liabilities, MarginAccount } from '../../data/MarginAccount';
