+++ prime/src/data/actions/ActionID.ts
@@ -1,0 +1,37 @@
+export enum ActionID {
+  TRANSFER_IN,
+  TRANSFER_OUT,
+  MINT,
+  BURN,
+  BORROW,
+  REPAY,
+  ADD_LIQUIDITY,
+  REMOVE_LIQUIDITY,
+  CLAIM_FEES,
+  SWAP,
+}
+
+export function getNameOfAction(id: ActionID): string {
+  switch (id) {
+    case ActionID.TRANSFER_IN:
+      return 'Add Margin';
+    case ActionID.TRANSFER_OUT:
+      return 'Withdraw';
+    case ActionID.MINT:
+      return 'Mint Token+';
+    case ActionID.BURN:
+      return 'Burn Token+';
+    case ActionID.BORROW:
+      return 'Borrow';
+    case ActionID.REPAY:
+      return 'Repay';
+    case ActionID.ADD_LIQUIDITY:
+      return 'Add Liquidity';
+    case ActionID.REMOVE_LIQUIDITY:
+      return 'Remove Liquidity';
+    case ActionID.CLAIM_FEES:
+      return 'Claim Fees';
+    default:
+      return 'UNKNOWN';
+  }
+}
