+++ prime/src/pages/BorrowActionsPage.tsx
@@ -32,7 +32,7 @@ import {
   calculateHypotheticalStates,
   UniswapPosition,
   UniswapPositionPrior,
-} from '../data/Actions';
+} from '../data/actions/Actions';
 import { ALOE_II_MARGIN_ACCOUNT_LENS_ADDRESS } from '../data/constants/Addresses';
 import { RESPONSIVE_BREAKPOINT_MD, RESPONSIVE_BREAKPOINT_XS } from '../data/constants/Breakpoints';
 import { useDebouncedEffect } from '../data/hooks/UseDebouncedEffect';
